% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoploss.R
\name{add.stoplossrule}
\alias{add.stoplossrule}
\title{add a stop loss rule to a strategy}
\usage{
add.stoplossrule(strategy, name, arguments, parameters = NULL, label = NULL,
  parent = NULL, type = "exit", ..., enabled = TRUE, indexnum = NULL,
  path.dep = TRUE, timespan = NULL, store = FALSE, storefun = TRUE)
}
\arguments{
\item{strategy}{an object of type 'strategy' to add the rule to}

\item{name}{name of the stop loss rule, must correspond to an R function}

\item{arguments}{named list of default arguments to be passed to an rule function when executed}

\item{parameters}{vector of strings naming parameters to be saved for apply-time definition}

\item{label}{arbitrary text label for rule output, NULL default will be converted to '<name>.rule'}

\item{parent}{the label of the parent rule for a chain rule, unavailable}

\item{type}{should be exit, and it is default}

\item{...}{any other passthru parameters}

\item{enabled}{TRUE/FALSE whether the rule is enabled for use in applying the strategy, default TRUE}

\item{indexnum}{if you are updating a specific rule, the index number in the $rules[type] list to update}

\item{path.dep}{TRUE/FALSE whether rule is path dependent, default TRUE, see Details}

\item{timespan}{an xts/ISO-8601 style \emph{time} subset, like "T08:00/T15:00", see Details}

\item{store}{TRUE/FALSE whether to store the strategy in the .strategy environment, or return it.  default FALSE}

\item{storefun}{TRUE/FALSE whether to store the function in the rule, default TRUE.  setting this option to FALSE may slow the backtest, but makes \code{\link{debug}} usable}
}
\value{
if \code{strategy} was the name of a strategy, the name. It it was a strategy, the updated strategy.
}
\description{
add a stop loss rule to a strategy
}

