% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoploss.R
\name{add.stoplosssig}
\alias{add.stoplosssig}
\title{add a stop loss signal to a strategy}
\usage{
add.stoplosssig(strategy, sigcol, arguments, parameters = NULL,
  label = NULL, type = c(NULL, "stopLimit", "stopPercent", "stopVolatility",
  "stopTargetperiod"), ..., enabled = TRUE, indexnum = NULL,
  store = FALSE)
}
\arguments{
\item{strategy}{an object of type 'strategy' to add the rule to}

\item{sigcol}{name of the trade signal, must correspond to the signal lable}

\item{arguments}{named list of default arguments to be passed to an stop loss rule function when executed}

\item{parameters}{vector of strings naming parameters to be saved for apply-time definition}

\item{label}{arbitrary text label for stop loss rule output, NULL default will be converted to '<name>.stop'}

\item{type}{one of "stopLimit","stopPercent","stopVolatility","stopTargetperiod" ,for now "stopVolatility" is unavailable}

\item{...}{any other passthru parameters}

\item{enabled}{TRUE/FALSE whether the rule is enabled for use in applying the strategy, default TRUE}

\item{indexnum}{if you are updating a specific rule, the index number in the $rules[type] list to update}

\item{store}{TRUE/FALSE whether to store the strategy in the .strategy environment, or return it.  default FALSE}
}
\value{
if \code{strategy} was the name of a strategy, the name. It it was a strategy, the updated strategy.
}
\description{
Actually, it's a special kind rules build on normal rule function.  
It will apply after  \code{\link{applyStoplossSig}}  
and add stop loss rules in the rule space.
}
\details{
All stop rules depends on the trade signals.
}

