% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoploss.R
\name{applyStoplossSig}
\alias{applyStoplossSig}
\title{apply the stop loss signal in the strategy to arbitrary market data}
\usage{
applyStoplossSig(strategy, mktdata, parameters = NULL, ..., enabled = TRUE)
}
\arguments{
\item{strategy}{an object of type 'strategy' to add the indicator to}

\item{mktdata}{an xts object containing market data.  depending on indicators, may need to be in OHLCV or BBO formats}

\item{parameters}{named list of parameters to be applied during evaluation of the strategy}

\item{...}{any other passthru parameters}

\item{enabled}{TRUE/FALSE whether the signal is enabled for use in applying the strategy, default TRUE}
}
\value{
\code{mktdata} with stop loss signal colums added.
}
\description{
\code{applyStoplossSig} will take the \code{mktdata} object, 
and will apply each stop loss signal specified in the strategy definition to it.
}

