% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoploss.R
\name{stopLimit}
\alias{stopLimit}
\title{generate a limit stop loss signal}
\usage{
stopLimit(sigcol, label, data = mktdata, column, limit, type = c("static",
  "dynamic"), orderside = c("long", "short"), offset1 = 0, offset2 = 0,
  ...)
}
\arguments{
\item{sigcol}{the column name of a trade signal}

\item{label}{text label to apply to the output}

\item{data}{data to apply}

\item{column}{named column to apply stop loss signal}

\item{limit}{limit loss of price}

\item{type}{one of c("static","dynamic")}

\item{orderside}{one of c("long","short") the orderside of the trade signal}

\item{offset1}{numeric offset to be added to the first column prior to comparison}

\item{offset2}{numeric offset to be added to the second column prior to comparison}

\item{...}{any other passthru parameters}
}
\description{
generate a limit stop loss signal
}

