% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoploss.R
\name{stopTargetperiod}
\alias{stopTargetperiod}
\title{generate a traget period stop loss signal}
\usage{
stopTargetperiod(sigcol, label, data = mktdata, period, ...)
}
\arguments{
\item{sigcol}{the column name of a trade signal}

\item{label}{text label to apply to the output}

\item{data}{data to apply}

\item{period}{target stop loss period}

\item{...}{any other passthru parameters}
}
\description{
generate a traget period stop loss signal
}

